Add more scripts from msf resources
and improved cscan.py to:
- Run a single script from args
- Run multiple scripts from args with a comma separated list
- Run default scripts (in config.py)
All these scripts are run in correct order as they are declared.
The command argument -p/--plugin was changed to -s/--script because
this argument is used to specify a script to run and not a
plugin.. Plugins are not intended to be ran directly from command line
but are used by scripts, which are used from cscan.py cli.
Also added -S/--scripts argument to pass a list of script to run.

Change in msfrpc.py to separate MSF options with ":" instead of
comma. Some options can contain some comma which are used to split
options list (example nmap --script args can contain comma for multiple
NSE scripts to use).

See msf-* scripts to have more example about how to use added scripts. A
quick example:
./cscan.py -S msf-basic-discovery-nmap.sh,msf-port-cleaner.sh,msf-auto-cred-checker.sh,msf-autocrawler.sh